+++ src/youwol/app/environment/models/predefined_configs/py_youwol_tour/new_ts_app.py
@@ -4,7 +4,7 @@ from youwol.app.environment import (
     Configuration,
     LocalEnvironment,
     Projects,
-    RecursiveProjectsFinder,
+    ProjectsFinder,
     System,
     get_standard_youwol_env,
 )
@@ -38,8 +38,8 @@ pipeline_ts.set_environment(
 
 Configuration(
     projects=Projects(
-        finder=RecursiveProjectsFinder(
-            fromPaths=[projects_folder],
+        finder=ProjectsFinder(
+            fromPath=projects_folder,
         ),
         templates=[app_ts_webpack_template(folder=projects_folder)],
     ),