+++ src/youwol/app/routers/projects/__init__.py
@@ -9,4 +9,4 @@ from .dependencies import *
 from .implementation import *
 from .models import *
 from .models_project import *
-from .projects_loader import *
+from .projects_resolver import *