+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [PEP 440 Versioning](https://peps.python.org/pep-044
 
 ### Changed
 
+- **Breaking:** Refactor the `ProjectsFinder` API and its associated implementation to improve performance and
+  flexibility. <!-- TG-2228 -->
 - Sanitize `EnvironmentStatusResponse` API from `admin/environment/status` endpoint. <!-- TG-2183 -->
 
 ### Fixed